langchain_version/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-qdrant>=0.0.1
qdrant-client>=1.7.0
openai>=1.30.0
python-dotenv==1.0.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langgraph==0.0.20
supabase==2.0.2
psycopg2-binary==2.9.7